Change for 1.3.0
