+++ viewbinding-nonreflection-ktx/src/main/java/com/dylanc/viewbinding/nonreflection/View.kt
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 2020. Dylan Cai
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@file:Suppress("unused")
+
+package com.dylanc.viewbinding.nonreflection
+
+import android.view.View
+import androidx.viewbinding.ViewBinding
+
+@Suppress("UNCHECKED_CAST")
+fun <VB : ViewBinding> View.getBinding(bind: (View) -> VB): VB =
+  (getTag(Int.MIN_VALUE) as? VB) ?: bind(this).also { setTag(Int.MIN_VALUE, it) }
