+++ viewbinding-nonreflection-ktx/src/main/java/com/dylanc/viewbinding/nonreflection/ViewGroup.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2020. Dylan Cai
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.dylanc.viewbinding.nonreflection
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.viewbinding.ViewBinding
+
+fun <VB : ViewBinding> ViewGroup.inflate(inflate: (LayoutInflater, ViewGroup?, Boolean) -> VB) =
+  inflate(LayoutInflater.from(context), this, true)
+
+fun <VB : ViewBinding> ViewGroup.binding(
+  inflate: (LayoutInflater, ViewGroup?, Boolean) -> VB,
+  attachToParent: Boolean = false
+) = lazy {
+  inflate(LayoutInflater.from(context), if (attachToParent) this else null, attachToParent)
+}
